Update Readme.md
Compare to branch "RISC-V", this branch "RISC-V-V2" is created based on the most recent edk2/master @37eef910. This is easier for reviewers to have clear ideas of edk2 code changes for RISC-V EDK2 implementation. Because of the code changes made on branch "RISC-V" is stale and not compliant with the latest RISC-V spec, this new branch has the fresh changes for RISC-V EDK2 implementation.